Fix cudart specs
